Track original framerate for video files
For certain media files ffmpeg made incorrect assumptions for the
framerate of the raw HEVC stream and this caused frames to be dropped
during transcoding and further A/V sync issues occurred in the muxed MP4
file. Adding these parameters makes sure that the file is kept at the
original framerate.